refactor weapon attacks and verbal combat
