+++ module/actor/sheets/mixins/itemMixin.js
@@ -1,19 +1,3 @@
-import { extendedRoll } from '../../../scripts/rolls/extendedRoll.js';
-
-import { rollDamage } from '../../../scripts/combat/attack.js';
-import { RollConfig } from '../../../scripts/rollConfig.js';
-import {
-    applyStatusEffectToActor,
-    applyStatusEffectToTargets
-} from '../../../scripts/statusEffects/applyStatusEffect.js';
-import { applyModifierToActor, applyModifierToTargets } from '../../../scripts/globalModifier/applyGlobalModifier.js';
-import {
-    applyActiveEffectToActor,
-    applyActiveEffectToTargets
-} from '../../../scripts/activeEffects/applyActiveEffect.js';
-
-const DialogV2 = foundry.applications.api.DialogV2;
-
 export let itemMixin = {
     async _onDropItem(event, data) {
         if (!this.actor.isOwner) return false;
@@ -254,387 +238,8 @@ export let itemMixin = {
         editor.toggleClass('invisible');
     },
 
-    createBaseDamageObject(item) {
-        return {
-            damageProperties: item.system.damageProperties,
-            item: item,
-            itemUuid: item.uuid,
-            crit: {
-                critLocationModifier: item.parent.system.attackStats.critLocationModifier,
-                critEffectModifier: item.parent.system.attackStats.critEffectModifier
-            }
-        };
-    },
-
     async _onItemRoll(event, itemId = null) {
-        let displayRollDetails = game.settings.get('TheWitcherTRPG', 'displayRollsDetails');
-
-        if (!itemId) {
-            itemId = event.currentTarget.closest('.item').dataset.itemId;
-        }
-        let item = this.actor.items.get(itemId);
-        let displayDmgFormula = `${item.system.damage}`;
-        let formula = !displayRollDetails
-            ? `${item.system.damage}`
-            : `${item.system.damage}[${game.i18n.localize('WITCHER.Diagram.Weapon')}]`;
-
-        let isMeleeAttack = item.doesWeaponNeedMeleeSkillToAttack();
-        if ((this.actor.type == 'character' || this.actor.system.addMeleeBonus) && isMeleeAttack) {
-            if (this.actor.system.attackStats.meleeBonus < 0) {
-                displayDmgFormula += `${this.actor.system.attackStats.meleeBonus}`;
-                formula += !displayRollDetails
-                    ? `${this.actor.system.attackStats.meleeBonus}`
-                    : `${this.actor.system.attackStats.meleeBonus}[${game.i18n.localize('WITCHER.Dialog.attackMeleeBonus')}]`;
-            }
-            if (this.actor.system.attackStats.meleeBonus > 0) {
-                displayDmgFormula += `+${this.actor.system.attackStats.meleeBonus}`;
-                formula += !displayRollDetails
-                    ? `+${this.actor.system.attackStats.meleeBonus}`
-                    : `+${this.actor.system.attackStats.meleeBonus}[${game.i18n.localize('WITCHER.Dialog.attackMeleeBonus')}]`;
-            }
-            formula = this.handleSpecialModifier(formula, 'melee-damage');
-        }
-
-        let attackSkill = item.getItemAttackSkill();
-        let messageData = {
-            speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-            flavor: `<h1> ${game.i18n.localize('WITCHER.Dialog.attack')}: ${item.name}</h1>`
-        };
-
-        let ammunitions = ``;
-        let noAmmo = 0;
-        let ammunitionOption = ``;
-        if (item.system.usingAmmo) {
-            ammunitions = this.actor.items.filter(function (item) {
-                return item.type == 'weapon' && item.system.isAmmo;
-            });
-            let quantity = ammunitions.sum('quantity');
-            if (quantity <= 0) {
-                noAmmo = 1;
-            } else {
-                ammunitions.forEach(element => {
-                    ammunitionOption += `<option value="${element._id}"> ${element.name}(${element.system.quantity}) </option>`;
-                });
-            }
-        }
-
-        let noThrowable = !this.actor.isEnoughThrowableWeapon(item);
-        let meleeBonus = this.actor.system.attackStats.meleeBonus;
-        let data = {
-            item,
-            attackSkill,
-            displayDmgFormula,
-            isMeleeAttack,
-            noAmmo,
-            noThrowable,
-            ammunitionOption,
-            ammunitions,
-            meleeBonus: meleeBonus
-        };
-        const myDialogOptions = { width: 500 };
-        const dialogTemplate = await renderTemplate(
-            'systems/TheWitcherTRPG/templates/dialog/combat/weapon-attack.hbs',
-            data
-        );
-
-        new Dialog(
-            {
-                title: `${game.i18n.localize('WITCHER.Dialog.attackWith')}: ${item.name}`,
-                content: dialogTemplate,
-                buttons: {
-                    Roll: {
-                        label: `${game.i18n.localize('WITCHER.Dialog.ButtonRoll')}`,
-                        callback: async html => {
-                            let isExtraAttack = html.find('[name=isExtraAttack]').prop('checked');
-
-                            let location = html.find('[name=location]')[0].value;
-                            let ammunition = undefined;
-                            if (html.find('[name=ammunition]')[0]) {
-                                ammunition = html.find('[name=ammunition]')[0].value;
-                            }
-
-                            let targetOutsideLOS = html.find('[name=targetOutsideLOS]').prop('checked');
-                            let outsideLOS = html.find('[name=outsideLOS]').prop('checked');
-                            let isFastDraw = html.find('[name=isFastDraw]').prop('checked');
-                            let isProne = html.find('[name=isProne]').prop('checked');
-                            let isPinned = html.find('[name=isPinned]').prop('checked');
-                            let isActivelyDodging = html.find('[name=isActivelyDodging]').prop('checked');
-                            let isMoving = html.find('[name=isMoving]').prop('checked');
-                            let isAmbush = html.find('[name=isAmbush]').prop('checked');
-                            let isRicochet = html.find('[name=isRicochet]').prop('checked');
-                            let isBlinded = html.find('[name=isBlinded]').prop('checked');
-                            let isSilhouetted = html.find('[name=isSilhouetted]').prop('checked');
-                            let customAim = html.find('[name=customAim]')[0].value;
-
-                            let range = item.system.range ? html.find('[name=range]')[0].value : null;
-                            let customAtt = html.find('[name=customAtt]')[0].value;
-                            let strike = html.find('[name=strike]')[0].value;
-                            let damageType = html.find('[name=damageType]')[0].value;
-                            let customDmg = html.find('[name=customDmg]')[0].value;
-                            let attacknumber = 1;
-
-                            let damage = this.createBaseDamageObject(item);
-                            damage = {
-                                ...damage,
-                                strike: strike,
-                                type: damageType
-                            };
-
-                            if (isExtraAttack) {
-                                let newSta = this.actor.system.derivedStats.sta.value - 3;
-
-                                if (newSta < 0) {
-                                    return ui.notifications.error(game.i18n.localize('WITCHER.Spell.notEnoughSta'));
-                                }
-                                this.actor.update({
-                                    'system.derivedStats.sta.value': newSta
-                                });
-                            }
-
-                            let allEffects = foundry.utils.deepClone(item.system.damageProperties.effects);
-                            if (ammunition) {
-                                let item = this.actor.items.get(ammunition);
-                                let newQuantity = item.system.quantity - 1;
-                                item.update({ 'system.quantity': newQuantity });
-                                allEffects.push(...item.system.damageProperties.effects);
-                                damage.ammunition = item;
-                            }
-
-                            if (item.isWeaponThrowable()) {
-                                let newQuantity = item.system.quantity - 1;
-                                if (newQuantity < 0) {
-                                    return;
-                                }
-                                item.update({ 'system.quantity': newQuantity });
-                            }
-
-                            if (item.system.enhancementItems) {
-                                item.system.enhancementItems.forEach(element => {
-                                    if (element && JSON.stringify(element) != '{}') {
-                                        let enhancement = this.actor.items.get(element.id);
-                                        allEffects.push(...enhancement.system.effects);
-                                    }
-                                });
-                            }
-                            damage.damageProperties.effects = allEffects;
-
-                            if (strike == 'fast') {
-                                attacknumber = 2;
-                            }
-                            for (let i = 0; i < attacknumber; i++) {
-                                let attFormula = '1d10+';
-                                let skill = CONFIG.WITCHER.skillMap[attackSkill.name];
-                                if (game.settings.get('TheWitcherTRPG', 'woundsAffectSkillBase')) {
-                                    attFormula += '(';
-                                }
-                                attFormula += !displayRollDetails
-                                    ? `${this.actor.system.stats[skill.attribute.name].current}+${this.actor.system.skills[skill.attribute.name][skill.name].value}`
-                                    : `${this.actor.system.stats[skill.attribute.name].current}[${game.i18n.localize(skill.attribute.label)}]+${this.actor.system.skills[skill.attribute.name][skill.name].value}[${game.i18n.localize(skill.label)}]`;
-
-                                attFormula = this.handleSpecialModifier(attFormula, strike);
-                                attFormula += this.actor.addAllModifiers(attackSkill.name);
-                                let damageFormula = formula;
-
-                                if (item.system.accuracy < 0) {
-                                    attFormula += !displayRollDetails
-                                        ? `${item.system.accuracy}`
-                                        : `${item.system.accuracy}[${game.i18n.localize('WITCHER.Weapon.Short.WeaponAccuracy')}]`;
-                                }
-                                if (item.system.accuracy > 0) {
-                                    attFormula += !displayRollDetails
-                                        ? `+${item.system.accuracy}`
-                                        : `+${item.system.accuracy}[${game.i18n.localize('WITCHER.Weapon.Short.WeaponAccuracy')}]`;
-                                }
-                                if (targetOutsideLOS) {
-                                    attFormula += !displayRollDetails
-                                        ? `-3`
-                                        : `-3[${game.i18n.localize('WITCHER.Dialog.attackTargetOutsideLOS')}]`;
-                                }
-                                if (outsideLOS) {
-                                    attFormula += !displayRollDetails
-                                        ? `+3`
-                                        : `+3[${game.i18n.localize('WITCHER.Dialog.attackOutsideLOS')}]`;
-                                }
-                                if (isExtraAttack) {
-                                    attFormula += !displayRollDetails
-                                        ? `-3`
-                                        : `-3[${game.i18n.localize('WITCHER.Dialog.attackExtra')}]`;
-                                }
-                                if (isFastDraw) {
-                                    attFormula += !displayRollDetails
-                                        ? `-3`
-                                        : `-3[${game.i18n.localize('WITCHER.Dialog.attackIsFastDraw')}]`;
-                                }
-                                if (isProne) {
-                                    attFormula += !displayRollDetails
-                                        ? `-2`
-                                        : `-2[${game.i18n.localize('WITCHER.Dialog.attackIsProne')}]`;
-                                }
-                                if (isPinned) {
-                                    attFormula += !displayRollDetails
-                                        ? `+4`
-                                        : `+4[${game.i18n.localize('WITCHER.Dialog.attackIsPinned')}]`;
-                                }
-                                if (isActivelyDodging) {
-                                    attFormula += !displayRollDetails
-                                        ? `-2`
-                                        : `-2[${game.i18n.localize('WITCHER.Dialog.attackIsActivelyDodging')}]`;
-                                }
-                                if (isMoving) {
-                                    attFormula += !displayRollDetails
-                                        ? `-3`
-                                        : `-3[${game.i18n.localize('WITCHER.Dialog.attackIsMoving')}]`;
-                                }
-                                if (isAmbush) {
-                                    attFormula += !displayRollDetails
-                                        ? `+5`
-                                        : `+5[${game.i18n.localize('WITCHER.Dialog.attackIsAmbush')}]`;
-                                }
-                                if (isRicochet) {
-                                    attFormula += !displayRollDetails
-                                        ? `-5`
-                                        : `-5[${game.i18n.localize('WITCHER.Dialog.attackIsRicochet')}]`;
-                                }
-                                if (isBlinded) {
-                                    attFormula += !displayRollDetails
-                                        ? `-3`
-                                        : `-3[${game.i18n.localize('WITCHER.Dialog.attackIsBlinded')}]`;
-                                }
-                                if (isSilhouetted) {
-                                    attFormula += !displayRollDetails
-                                        ? `+2`
-                                        : `+2[${game.i18n.localize('WITCHER.Dialog.attackIsSilhouetted')}]`;
-                                }
-                                if (customAim > 0) {
-                                    attFormula += !displayRollDetails
-                                        ? `+${customAim}`
-                                        : `+${customAim}[${game.i18n.localize('WITCHER.Dialog.attackCustom')}]`;
-                                }
-
-                                if (customAtt != '0') {
-                                    attFormula += !displayRollDetails
-                                        ? `+${customAtt}`
-                                        : `+${customAtt}[${game.i18n.localize('WITCHER.Settings.Custom')}]`;
-                                }
-
-                                switch (range) {
-                                    case 'pointBlank':
-                                        attFormula = !displayRollDetails
-                                            ? `${attFormula}+5`
-                                            : `${attFormula} +5[${game.i18n.localize('WITCHER.Weapon.Range')}]`;
-                                        break;
-                                    case 'medium':
-                                        attFormula = !displayRollDetails
-                                            ? `${attFormula}-2`
-                                            : `${attFormula} -2[${game.i18n.localize('WITCHER.Weapon.Range')}]`;
-                                        break;
-                                    case 'long':
-                                        attFormula = !displayRollDetails
-                                            ? `${attFormula}-4`
-                                            : `${attFormula} -4[${game.i18n.localize('WITCHER.Weapon.Range')}]`;
-                                        break;
-                                    case 'extreme':
-                                        attFormula = !displayRollDetails
-                                            ? `${attFormula}-6`
-                                            : `${attFormula} -6[${game.i18n.localize('WITCHER.Weapon.Range')}]`;
-                                        break;
-                                }
-
-                                if (customDmg != '0') {
-                                    damageFormula += !displayRollDetails
-                                        ? `+${customDmg}`
-                                        : `+${customDmg}[${game.i18n.localize('WITCHER.Settings.Custom')}]`;
-                                }
-                                damage.formula = damageFormula;
-
-                                let touchedLocation = this.actor.getLocationObject(location);
-                                attFormula += !displayRollDetails
-                                    ? `${touchedLocation.modifier}`
-                                    : `${touchedLocation.modifier}[${touchedLocation.alias}]`;
-                                damage.location = touchedLocation;
-                                damage.originalLocation = location;
-
-                                if (strike == 'joint') {
-                                    attFormula = !displayRollDetails
-                                        ? `${attFormula} -3${
-                                              this.actor.system.lifepathModifiers.jointStrikeAttackBonus > 0
-                                                  ? ` +${this.actor.system.lifepathModifiers.jointStrikeAttackBonus}`
-                                                  : ''
-                                          }`
-                                        : `${attFormula} -3[${game.i18n.localize('WITCHER.Dialog.attackStrike')}]${
-                                              this.actor.system.lifepathModifiers.jointStrikeAttackBonus > 0
-                                                  ? ` +${this.actor.system.lifepathModifiers.jointStrikeAttackBonus}[${game.i18n.localize('WITCHER.Actor.Lifepath.Bonus')}]`
-                                                  : ''
-                                          }`;
-                                }
-
-                                if (strike == 'strong') {
-                                    if (!displayRollDetails) {
-                                        attFormula = `${attFormula} -3${
-                                            this.actor.system.lifepathModifiers.strongStrikeAttackBonus > 0
-                                                ? ` +${this.actor.system.lifepathModifiers.strongStrikeAttackBonus}`
-                                                : ''
-                                        }`;
-                                    } else {
-                                        attFormula = `${attFormula} -3[${game.i18n.localize('WITCHER.Dialog.attackStrike')}]${
-                                            this.actor.system.lifepathModifiers.strongStrikeAttackBonus > 0
-                                                ? ` +${this.actor.system.lifepathModifiers.strongStrikeAttackBonus}[${game.i18n.localize('WITCHER.Actor.Lifepath.Bonus')}]`
-                                                : ''
-                                        }`;
-                                    }
-                                }
-
-                                messageData.flavor = `<div class="attack-message"><h1><img src="${item.img}" class="item-img" />${game.i18n.localize('WITCHER.Attack')}: ${item.name}</h1>`;
-                                messageData.flavor += `<span>  ${game.i18n.localize('WITCHER.Armor.Location')}: ${touchedLocation.alias} </span>`;
-
-                                messageData.flavor += `<button class="damage">${game.i18n.localize('WITCHER.table.Damage')}</button>`;
-                                if (item.system.rollOnlyDmg) {
-                                    rollDamage(item, damage);
-                                } else {
-                                    messageData.flags = {
-                                        TheWitcherTRPG: {
-                                            attack: item.getAttackSkillFlags(),
-                                            damage: damage
-                                        }
-                                    };
-                                    await extendedRoll(attFormula, messageData, new RollConfig());
-                                }
-                            }
-                        }
-                    }
-                }
-            },
-            myDialogOptions
-        ).render(true);
-    },
-
-    handleSpecialModifier(attFormula, action, additionalTag) {
-        let relevantModifier = this.actor
-            .getList('globalModifier')
-            .filter(modifier => modifier.system.isActive)
-            .filter(modifier => modifier.system.special?.length > 0)
-            .map(modifier => modifier.system.special)
-            .flat()
-            .map(modifier => CONFIG.WITCHER.specialModifier.find(special => special.id == modifier.special))
-            .filter(special => special?.tags?.includes(action))
-            .filter(special => special?.additionalTags?.includes(additionalTag?.toLowerCase()) ?? true);
-
-        let relevantActorModifier = this.actor.system.specialSkillModifiers
-            .map(specialSkillModifier =>
-                CONFIG.WITCHER.specialModifier.find(special => special.id == specialSkillModifier.modifier)
-            )
-            .filter(special => special?.tags?.includes(action))
-            .filter(special => special?.additionalTags?.includes(additionalTag?.toLowerCase()) ?? true);
-
-        relevantModifier
-            .concat(relevantActorModifier)
-            .forEach(modifier => (attFormula += `${modifier.formula}[${game.i18n.localize(modifier.name)}]`));
-
-        return attFormula;
-    },
-
-    async _onSpellRoll(event) {
-        this.actor.useItem(event.currentTarget.closest('.item').dataset.itemId);
+        this.actor.useItem(itemId ?? event.currentTarget.closest('.item').dataset.itemId);
     },
 
     _onSpellDisplay(event) {
@@ -675,6 +280,6 @@ export let itemMixin = {
         html.find('.spell-display').on('click', this._onSpellDisplay.bind(this));
 
         html.find('.item-roll').on('click', this._onItemRoll.bind(this));
-        html.find('.spell-roll').on('click', this._onSpellRoll.bind(this));
+        html.find('.spell-roll').on('click', this._onItemRoll.bind(this));
     }
 };
